requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.3
apturl==0.5.2
argilla==1.6.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.22
cmake==3.26.3
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
commonmark==0.9.1
contourpy==1.1.0
cryptography==40.0.2
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.7
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.2
dnspython==2.4.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duckdb==0.7.1
duplicity==0.8.12.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
Faker==19.1.0
fastapi==0.95.1
fasteners==0.14.1
favicon==0.7.0
filelock==3.12.0
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httpcore==0.16.3
httplib2==0.14.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.0
ipykernel==6.22.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.4.4
langchain==0.0.245
langsmith==0.0.15
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.2
lockfile==0.12.2
loguru==0.7.0
louis==3.12.0
lxml==4.9.2
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==4.2.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
paramiko==2.6.0
parso==0.8.3
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==11.9.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.17
pycups==1.9.73
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymdown-extensions==10.1
Pympler==1.0.1
PyNaCl==1.3.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
reportlab==3.5.34
requests==2.29.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==2.3.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.16.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlparse==0.4.4
st-annotated-text==4.0.0
stack-data==0.6.2
starlette==0.26.1
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.11.1
systemd-python==234
tenacity==8.2.2
texttable==1.6.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unstructured==0.6.2
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==0.53.0
websockets==11.0.2
wrapt==1.14.1
xkit==0.0.0
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
